Moving all implicit definition to one file rather than keeping them in SnappyContext
